add sample centring methods and some signals in Genericdiffractometer
