+++ GenericDiffractometer.py
@@ -25,7 +25,8 @@ import copy
 import time
 import gevent
 import logging
-
+import sample_centring
+import math,numpy
 import queue_model_objects_v1 as queue_model_objects
 
 from HardwareRepository.TaskUtils import *
@@ -169,6 +170,9 @@ class GenericDiffractometer(HardwareObject):
         self.command_dict = {}
         self.used_commands_list = GenericDiffractometer.COMMANDS_NAME
 
+	# flag for using sample_centring hwobj or not
+	self.use_sample_centring = None 
+
         # Internal values -----------------------------------------------------
         self.ready_event = None
         self.head_type = GenericDiffractometer.HEAD_TYPE_MINIKAPPA
@@ -265,6 +269,9 @@ class GenericDiffractometer(HardwareObject):
                 self.connect(self.channel_dict["CurrentPhase"], "update", self.current_phase_changed)
 	    elif channel_name =="HeadType":
                 self.connect(self.channel_dict["HeadType"], "update", self.head_type_changed)
+	    elif channel_name =="State":
+                self.connect(self.channel_dict["State"], "update", self.state_changed)
+
 
         # config motors from xml -----------------------------------------------------
         try:
@@ -282,6 +289,36 @@ class GenericDiffractometer(HardwareObject):
             if self.transfer_mode is None or self.transfer_mode == "SAMPLE_CHANGER":
                 self.use_sc = True
 
+
+        # centring motors
+        self.centring_motors_list = ["phiz", "phiy","sampx","sampy"]
+	try:
+           self.use_sample_centring = self.getProperty("sample_centring")
+	   if self.use_sample_centring:
+		self.centring_phi=sample_centring.CentringMotor(self.motor_hwobj_dict['phi'], direction=-1)
+	        self.centring_phiz=sample_centring.CentringMotor(self.motor_hwobj_dict['phiz'])
+        	self.centring_phiy=sample_centring.CentringMotor(self.motor_hwobj_dict['phiy'],direction=-1)
+        	self.centring_sampx=sample_centring.CentringMotor(self.motor_hwobj_dict['sampx'])
+        	self.centring_sampy=sample_centring.CentringMotor(self.motor_hwobj_dict['sampy'])
+        except:
+           pass # used the default value
+        
+
+        for motor_name in self.centring_motors_list:
+            if self.motor_hwobj_dict[motor_name] is not None:
+                self.connect(self.motor_hwobj_dict[motor_name], 'stateChanged', self.motor_state_changed)
+                self.connect(self.motor_hwobj_dict[motor_name], "positionChanged", self.centring_motor_moved)
+            else:
+                logging.getLogger("HWR").warning('%s motor is not defined' % motor_name)
+
+        if self.motor_hwobj_dict["phi"] is not None:
+            self.connect(self.motor_hwobj_dict["phi"], 'stateChanged', self.motor_state_changed)
+            self.connect(self.motor_hwobj_dict["phi"], "positionChanged", self.emit_diffractometer_moved)
+        else:
+            logging.getLogger("HWR").error('Phi motor is not defined')
+
+	# other motors
+	# to add, zoom
 
         self.front_light_swtich = self.getObjectByRole('frontlightswtich')
         self.back_light_swtich = self.getObjectByRole('backlightswtich')
@@ -555,14 +592,35 @@ class GenericDiffractometer(HardwareObject):
         """
         """
         self.emit_progress_message("Manual 3 click centring...")
-        self.current_centring_procedure = gevent.spawn(self.manual_centring)
+	if self.use_sample_centring:
+            self.current_centring_procedure = sample_centring.start({"phi": self.centring_phi,
+                                                               "phiy": self.centring_phiy,
+                                                               "sampx": self.centring_sampx,
+                                                               "sampy": self.centring_sampy,
+                                                               "phiz": self.centring_phiz },
+                                                              self.pixels_per_mm_x, self.pixels_per_mm_y,
+                                                              self.beam_position[0], self.beam_position[1])
+	else:
+            self.current_centring_procedure = gevent.spawn(self.manual_centring)
         self.current_centring_procedure.link(self.centring_done)
 
     def start_automatic_centring(self, sample_info=None, loop_only=False, wait_result=None):
         """
         """
         self.emit_progress_message("Automatic centring...")
-        self.current_centring_procedure = gevent.spawn(self.automatic_centring)
+	if self.use_sample_centring:
+            self.current_centring_procedure = sample_centring.start_auto(self.camera_hwobj,
+                                                                     {"phi": self.centring_phi,
+                                                                      "phiy": self.centring_phiy,
+                                                                      "sampx": self.centring_sampx,
+                                                                      "sampy": self.centring_sampy,
+                                                                      "phiz": self.centring_phiz },
+                                                                     self.pixels_per_mm_x, self.pixels_per_mm_y,
+                                                                     self.beam_position[0], self.beam_position[1],
+                                                                     msg_cb=self.emit_progress_message,
+                                                                     new_point_cb=lambda point: self.emit("newAutomaticCentringPoint", point))
+	else:
+            self.current_centring_procedure = gevent.spawn(self.automatic_centring)
         self.current_centring_procedure.link(self.centring_done)
 
         if wait_result:
@@ -635,8 +693,41 @@ class GenericDiffractometer(HardwareObject):
     def automatic_centring(self):
         raise NotImplementedError
 
+    def centring_motor_moved(self, pos):
+        if time.time() - self.centring_time > 1.0:
+            self.invalidate_centring()
+        self.emit_diffractometer_moved()
+
+    def invalidate_centring(self):
+        if self.current_centring_procedure is None and self.centring_status["valid"]:
+            self.centring_status={"valid":False}
+            self.emit_progress_message("")
+            self.emit('centringInvalid', ())
+
+    def emit_diffractometer_moved(self, *args):
+        self.emit("diffractometerMoved", ())
+
     def motor_positions_to_screen(self, centred_positions_dict):
-        raise NotImplementedError
+        if self.use_sample_centring:
+            self.update_zoom_calibration()
+            if None in (self.pixels_per_mm_x, self.pixels_per_mm_y):
+                return 0,0
+            phi_angle = math.radians(self.centring_phi.direction*self.centring_phi.getPosition())
+            sampx = self.centring_sampx.direction * (centred_positions_dict["sampx"]-self.centring_sampx.getPosition())
+            sampy = self.centring_sampy.direction * (centred_positions_dict["sampy"]-self.centring_sampy.getPosition())
+            phiy = self.centring_phiy.direction * (centred_positions_dict["phiy"]-self.centring_phiy.getPosition())
+            phiz = self.centring_phiz.direction * (centred_positions_dict["phiz"]-self.centring_phiz.getPosition())
+            rot_matrix = numpy.matrix([math.cos(phi_angle), -math.sin(phi_angle), math.sin(phi_angle), math.cos(phi_angle)])
+            rot_matrix.shape = (2, 2)
+            inv_rot_matrix = numpy.array(rot_matrix.I)
+            dx, dy = numpy.dot(numpy.array([sampx, sampy]), inv_rot_matrix)*self.pixels_per_mm_x
+
+            x = (phiy * self.pixels_per_mm_x) + self.beam_position[0]
+            y = dy + (phiz * self.pixels_per_mm_y) + self.beam_position[1]
+
+            return x, y
+        else:
+            raise NotImplementedError
 
     def move_to_centred_position(self, centred_position):
         self.move_motors(centred_position)
@@ -691,7 +782,10 @@ class GenericDiffractometer(HardwareObject):
         """
         Descript. :
         """
-        self.user_clicked_event.set((x, y))
+	if self.use_sample_centring:
+   	    sample_centring.user_click(x,y)
+	else:
+            self.user_clicked_event.set((x, y))
 
     def accept_centring(self):
         """
@@ -843,6 +937,15 @@ class GenericDiffractometer(HardwareObject):
     def equipment_not_ready(self):
         self.emit('minidiffNotReady', ())
 
+    def state_changed(self, state):
+        self.current_state = state
+        self.emit("minidiffStateChanged", (self.current_state))
+
+    def motor_state_changed(self, state):
+        """
+        Descript. :
+        """
+        self.emit('minidiffStateChanged', (state, ))
 
     def current_phase_changed(self, current_phase):
         """
